Update docs, add license, rename src/examples -> src/doc_examples
