+++ src/doc_examples/tree.rs
